Add files via upload
added a hamburger menu on the navigation bar when viewed on smaller screens. added a sticky, fade-in nav bar that is visible once you scroll past the hero section. used css to style the nav bar and js to make sure that it works perfectly. separated the about me and contact links that was in the nav bar. the featured destination images were added and styled using css. there are three static images. two preview blog post were created along with a read more button that links to a different web page, that is coming soon. the layout of the page was changed slightly, there is now a side bar and a main content section so that the users will be able to view everything without having to scroll to read a little bit about myself or see the links to my accounts. links does not work as of yet.