Add files via upload
changed the button in the text block on hero background to an animated button. That will be changed, I don't like the arrow. I will remove the subtle shift on hover. On mobile device, it takes a little longer for the menu to slide out from the right side which is what I was going for.
Created the page to post blogs about my vacations and styled it using css and made sure that it is responsive.