Add files via upload
added a darker tint to the hero section. changed the wording on the button and styled it so that it is smaller.
added a slightly darker tint to the welcome section. moved the welcome message to the sidebar and replaced with different content
gave the bottom of the card a linear gradient color
the button in the header scrolls to most recent stories